+++ src/settings/MappingProfiles/MappingProfilesFormContainer/processRawTransformations.js
@@ -1,7 +1,12 @@
 import { omit } from 'lodash';
 
+import { isInstanceTransformationEmpty } from '../MappingProfilesTransformationsModal/validateTransformations';
+
 export const parseRawTransformation = transformation => {
-  // TODO: remove this check once transformations validation is done in scope of UIDEXP-187.
+  if (isInstanceTransformationEmpty(transformation)) {
+    return '';
+  }
+
   const { rawTransformation = {} } = transformation;
   const {
     marcField = '',
@@ -24,7 +29,7 @@ export const parseRawTransformation = transformation => {
 };
 
 export const splitIntoRawTransformation = transformation => {
-  const regex = /^(?<marcField>\d{3})(?<indicator1>[\d\s])(?<indicator2>[\d\s])(?<subfield>\$([0-9]{1,2}|[a-zA-Z]))?$/;
+  const regex = /^(?<marcField>\d{3})(?<indicator1>[\d\sa-zA-Z])(?<indicator2>[\d\sa-zA-Z])(?<subfield>\$([0-9]{1,2}|[a-zA-Z]))?$/;
   const rawTransformation = transformation.match(regex)?.groups;
 
   if (rawTransformation?.indicator1 === ' ') {